requirements.txt
Django==5.0.7
djangorestframework==3.16.0
drf-yasg==1.21.10
Pillow==10.4.0
python-decouple==3.8
psycopg2==2.9.10
PyYAML==6.0.2
uritemplate==4.1.1
inflection==0.5.1
djangorestframework-simplejwt==5.3.1 